% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{lines_to_cells}
\alias{lines_to_cells}
\title{Convert Lines To Cells}
\usage{
lines_to_cells(lines, tol_x = 5, tol_y = tol_x, plot = FALSE)
}
\arguments{
\item{lines}{(numeric) A matrix of line coordinates.}

\item{tol_x}{(numeric) The tolerance for joining lines along the x-axis.}

\item{tol_y}{(numeric) The tolerance for joining lines along the y-axis}

\item{plot}{(logical) If TRUE, plot cell detection information.}
}
\value{
A matrix of cell (rectangle) coordinates.
}
\description{
Convert Lines To Cells
}

