% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_geometry.R
\name{plot_pdf_lines}
\alias{plot_pdf_lines}
\title{Plot PDF Lines}
\usage{
plot_pdf_lines(x, lty = 1, col = 1, ...)
}
\arguments{
\item{x}{(numeric) A matrix of lines.}

\item{...}{Further arguments to the matlines function.}
}
\description{
Plot PDF Lines
}

