% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_geometry.R
\name{plot_pdf_new}
\alias{plot_pdf_new}
\title{New PDF Page Plot}
\usage{
plot_pdf_new(page, ...)
}
\arguments{
\item{page}{A page node from an XML-converted PDF.}
}
\description{
This function opens a blank plot sized for a PDF page.
}

