% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_geometry.R
\name{plot_pdf_page}
\alias{plot_pdf_page}
\title{Plot PDF Page Geometry}
\usage{
plot_pdf_page(page)
}
\arguments{
\item{page}{A page node from an XML-converted PDF.}
}
\description{
This function plots the geometry in a page node from an XML-converted PDF.
}

