% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_geometry.R
\name{plot_pdf_rects}
\alias{plot_pdf_rects}
\title{Plot PDF Rects}
\usage{
plot_pdf_rects(x, ...)
}
\arguments{
\item{x}{(numeric) A matrix of rects.}

\item{...}{Further arguments to the rect function.}
}
\description{
Plot PDF Rects
}

