% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{rects_to_lines}
\alias{rects_to_lines}
\title{Convert Rects To Lines}
\usage{
rects_to_lines(rects)
}
\arguments{
\item{x}{(numeric) A matrix of rect coordinates.}
}
\value{
(numeric) A matrix of line coordinates.
}
\description{
Convert Rects To Lines
}

