% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{simplify_lines_hz}
\alias{simplify_lines_hz}
\title{Simplify Horizontal Lines}
\usage{
simplify_lines_hz(lines, tol_line = 0, tol_join = 0)
}
\arguments{
\item{lines}{(numeric) A matrix of horizontal line coordinates.}

\item{tol_line}{(numeric) The tolerance for determining colinearity.}

\item{tol_join}{(numeric) The tolerance for joining colinear segments.}
}
\value{
(list) A list of matrices, one for each set of colinear segments.
}
\description{
Simplify Horizontal Lines
}

