% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{split_lines_hv}
\alias{split_lines_hv}
\title{Split Lines Into Horizontals & Verticals}
\usage{
split_lines_hv(lines, tol = 0)
}
\arguments{
\item{(numeric)}{A matrix of line coordinates.}
}
\value{
(list) A list which contains a matrix of horizontal lines and a
matrix of vertical lines.
}
\description{
Split Lines Into Horizontals & Verticals
}

